feat: port binding improvements
 - add `util::new_random_tcp_listener`.
 - add `util::new_random_tcp_listener_with_socket_addr`.
 - `TestServer` requests a `TcpListener` over a reserved port from
   `reserved-port`.

This is a part of an effort in locking down the port binding behaviour.
To guarantee two test servers bind the same port.

This includes adding tools to make it easier to build your own
test servers (such as for mocking), which will run externally to this,
and don't clash on the port.